compute euclidian distance between matching points

+++ following/displacement_vector.h
@@ -9,9 +9,9 @@
 #include <cv.h>
 #include <pthread.h>
 
-int cross_correlation(uchar** image_1, uchar** image_2, int size_x, int size_y, int ptx1, int ptx2, int pty1, int pty2, int size_win_x, int size_win_y);
+double cross_correlation(uchar** image_1, uchar** image_2, int size_x, int size_y, int ptx1, int ptx2, int pty1, int pty2, int size_win_x, int size_win_y);
 int matching_point(uchar** image_1 , uchar** image_2, int size_x, int size_y, int ptx, int pty, uchar** harris2, int window_size);
-int euclidian_distance(uchar** image_1, uchar** image_2, int size_x, int size_y, int ptx1, int ptx2, int pty1, int pty2, int size_win_x, int size_win_y);
+double euclidian_distance(uchar** image_1, uchar** image_2, int size_x, int size_y, int ptx1, int ptx2, int pty1, int pty2, int size_win_x, int size_win_y);
